python-cdktr/pyproject.toml
[project]
name = "python-cdktr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyzmq>=26.2.0",
    "rustyrs>=0.5.5",
]
